remove tests from benchmarks
