+++ word2vec_ziutek.go
